[refs #36160] Refactor enterprises tests
