+++ core/api/filters/enterprise.py
@@ -1,25 +1,10 @@
-from django.db.models import Q
 from django_filters import rest_framework as filters
 from django_filters.fields import CSVWidget
-from django_filters.rest_framework import DjangoFilterBackend
 
 from core.models.agency import Agency
 from core.models.country import Country
 from core.models.meeting import Meeting
-from core.models.project import (
-    MetaProject,
-    Project,
-)
-from core.models.enterprise import Enterprise, EnterpriseStatus
-from core.models.project_metadata import (
-    ProjectCluster,
-    ProjectSector,
-    ProjectStatus,
-    ProjectSubmissionStatus,
-    ProjectSubSector,
-    ProjectType,
-)
-from core.models.utils import SubstancesType
+from core.models.enterprise import Enterprise
 
 
 class EnterpriseFilter(filters.FilterSet):
