+++ core/import_data_v2/resources/users/permissions.json
@@ -222,24 +222,6 @@
         "model_name": "enterprise"
     },
     {
-        "permission_codename": "has_enterprise_approval_access",
-        "permission_name": "Has enterprise approval access",
-        "app_label": "core",
-        "model_name": "enterprise"
-    },
-    {
-        "permission_codename": "has_project_enterprise_edit_access",
-        "permission_name": "Has project enterprise edit access",
-        "app_label": "core",
-        "model_name": "projectenterprise"
-    },
-    {
-        "permission_codename": "has_project_enterprise_approval_access",
-        "permission_name": "Has project enterprise approval access",
-        "app_label": "core",
-        "model_name": "projectenterprise"
-    },
-    {
         "permission_codename": "has_project_settings_access",
         "permission_name": "Has project settings view and edit access",
         "app_label": "core",
